refactor data of subgraph response
